templ: move to klplib package
